requirements/base.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
environs[django]==9.5.0
parameterized==0.8.1
djangorestframework-simplejwt==5.2.2